requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
streamlit==1.12.2

